requirements.txt
alembic==0.9.1
appdirs==1.4.3
click==6.7
Flask==0.12.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
markdown2==2.3.3
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-editor==1.0.3
python-slugify==1.2.4
six==1.10.0
SQLAlchemy==1.1.9
Unidecode==0.4.20
Werkzeug==0.12.1
